refactor: 과한 객체 분리를 제거하기 위한 MailContentWriter 제거
