+++ src/main/java/com/gdschongik/gdsc/domain/email/application/UnivEmailVerificationLinkSendService.java
@@ -5,7 +5,6 @@ import com.gdschongik.gdsc.domain.email.dao.UnivEmailVerificationRepository;
 import com.gdschongik.gdsc.domain.email.domain.UnivEmailVerification;
 import com.gdschongik.gdsc.domain.email.util.VerificationCodeGenerator;
 import com.gdschongik.gdsc.domain.email.util.VerificationLinkUtil;
-import com.gdschongik.gdsc.domain.email.util.VerificationMailContentWriter;
 import com.gdschongik.gdsc.domain.member.dao.MemberRepository;
 import com.gdschongik.gdsc.global.exception.CustomException;
 import com.gdschongik.gdsc.global.exception.ErrorCode;
@@ -29,16 +28,27 @@ public class UnivEmailVerificationLinkSendService {
     private final HongikUnivEmailValidator hongikUnivEmailValidator;
     private final VerificationCodeGenerator verificationCodeGenerator;
     private final VerificationLinkUtil verificationLinkUtil;
-    private final VerificationMailContentWriter verificationMailContentWriter;
     private final MemberUtil memberUtil;
     public static final Duration VERIFICATION_CODE_TIME_TO_LIVE = Duration.ofMinutes(10);
+
+    private static final String NOTIFICATION_MESSAGE = "<div style='margin:20px;'>"
+            + "<h1> 안녕하세요 GDSC Hongik 재학생 인증 메일입니다. </h1> <br>"
+            + "<h3> 아래의 링크를 %d분 안에 클릭해주세요. </h3> <br>"
+            + "<h3> 감사합니다. </h3> <br>"
+            + "CODE : <strong>";
+
+    public String write(String verificationLink, Duration codeAliveTime) {
+        return String.format(NOTIFICATION_MESSAGE, codeAliveTime.toMinutes()) + verificationLink;
+    }
 
     public void send(String univEmail) {
         hongikUnivEmailValidator.validate(univEmail);
         validateUnivEmailNotVerified(univEmail);
 
         String verificationCode = verificationCodeGenerator.generate();
-        sendVerificationLink(univEmail, verificationCode);
+        String verificationLink = verificationLinkUtil.createLink(verificationCode);
+        String mailContent = writeMailContentWithVerificationLink(verificationLink);
+        mailSender.send(univEmail, UnivEmailVerificationConstant.VERIFICATION_EMAIL_SUBJECT, mailContent);
 
         saveUnivEmailVerification(univEmail, verificationCode);
     }
@@ -49,11 +59,8 @@ public class UnivEmailVerificationLinkSendService {
         });
     }
 
-    private void sendVerificationLink(String univEmail, String verificationCode) {
-        String verificationLink = verificationLinkUtil.createLink(verificationCode);
-        String mailContent = verificationMailContentWriter.write(verificationLink, VERIFICATION_CODE_TIME_TO_LIVE);
-
-        mailSender.send(univEmail, UnivEmailVerificationConstant.VERIFICATION_EMAIL_SUBJECT, mailContent);
+    private String writeMailContentWithVerificationLink(String verificationLink) {
+        return String.format(NOTIFICATION_MESSAGE, VERIFICATION_CODE_TIME_TO_LIVE.toMinutes()) + verificationLink;
     }
 
     private void saveUnivEmailVerification(String univEmail, String verificationCode) {
